+++ trailer.c
@@ -299,7 +299,7 @@ static void apply_arg_if_exists(struct trailer_item *in_tok,
 		break;
 	default:
 		BUG("trailer.c: unhandled value %d",
-		    arg_tok->conf.if_exists);
+		    (int)arg_tok->conf.if_exists);
 	}
 }
 
@@ -324,7 +324,7 @@ static void apply_arg_if_missing(struct list_head *head,
 		break;
 	default:
 		BUG("trailer.c: unhandled value %d",
-		    arg_tok->conf.if_missing);
+		    (int)arg_tok->conf.if_missing);
 	}
 }
 
@@ -557,7 +557,7 @@ static int git_trailer_config(const char *conf_key, const char *value, void *cb)
 			warning(_("unknown value '%s' for key '%s'"), value, conf_key);
 		break;
 	default:
-		BUG("trailer.c: unhandled type %d", type);
+		BUG("trailer.c: unhandled type %d", (int)type);
 	}
 	return 0;
 }
